+++ src/components/avatar/avatar.tsx
@@ -109,6 +109,7 @@ export class Avatar {
       }
     };
   }
+
   render() {
     return (
       <>