+++ plus.config.js
@@ -13,6 +13,8 @@ import {
   visualStudioCode,
   webTypes,
 } from '@htmlplus/element/compiler/index.js';
+import fs from 'fs';
+import path from 'path';
 
 import PACKAGE from './package.json' assert { type: 'json' };
 
@@ -117,5 +119,41 @@ export default [
         ?.filter((line) => !!line.trim())[0]
         ?.trim();
     }
-  })
+  }),
+  {
+    name: 'theme',
+    finish: (global) => {
+      const outputs = {};
+
+      for (const context of global.contexts) {
+        try {
+          const directory = path.join(context.directoryPath, 'theme');
+
+          const files = fs.readdirSync(directory);
+
+          for (const file of files) {
+            const content = fs.readFileSync(path.join(directory, file), 'utf8');
+
+            outputs[`${file.replace('.css', '')}/${context.fileName}.css`] = content;
+
+            outputs[file] ||= '';
+
+            outputs[file] += content + '\n';
+          }
+        } catch {}
+      }
+      
+      for (const output in outputs) {
+        if (!Object.hasOwnProperty.call(outputs, output)) return;
+
+        const file = path.join('dist', 'theme', output);
+
+        const directory = path.dirname(file);
+        
+        if (!fs.existsSync(directory)) fs.mkdirSync(directory, { recursive: true }); 
+
+        fs.writeFileSync(file, outputs[output], 'utf8');
+      }
+    }
+  },
 ];