+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Updated
 
+- Improved the `Global Config`.
 - Renamed the `source` property to `reference` in the `Scroll Detector` component.
 - The error message for third-party components has been updated.
 - Updated the `outside click` dependencies.