+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the themeing system.
-- Added the experimental `override` property to the components.
+- The `accordion` component.
 
 ### Updated
 
+- Refactor the `icon`.
 - Moved the animation assets.
 - Added the `initializer` property for `styles` to the `json/document.json`.